Updated vegetable Class
